Moving dialog.display_message to utils
