+++ lib/python/node_manager/manager.py
@@ -22,7 +22,6 @@ from node_manager import utils
 from node_manager.utils import plugin
 from node_manager.utils import callbacks
 from node_manager.utils import definition as definitionutils
-from node_manager.dependencies import dialog
 from node_manager.dependencies import nodes
 
 logger = logging.getLogger(__name__)
@@ -586,7 +585,7 @@ class NodeManager(object):
             # repo.remove_definition(definition)
 
             # Success
-            dialog.display_message(
+            utils.display_message(
                 "HDA release successful!", title="HDA Manager: Publish HDA"
             )
